Implemented command pattern for if statements

+++ src/main/java/ByteRange.java
@@ -1,25 +1,25 @@
 public class ByteRange {
-    private int _start;
-    private int _end;
+    private Long _start;
+    private Long _end;
 
     public ByteRange() {
-        _start = 0;
-        _end = 0;
+        _start = 0L;
+        _end = 0L;
     }
 
-    public void setStart(int start) {
+    public void setStart(Long start) {
         _start = start;
     }
 
-    public void setEnd(int end) {
+    public void setEnd(Long end) {
         _end = end;
     }
 
-    public int get_start() {
+    public Long get_start() {
         return _start;
     }
 
-    public int get_end() {
+    public Long get_end() {
         return _end;
     }
 }
